requirements.txt
alpaca-trade-api
python-dotenv